add translate_note_to_freq in test_script, tested correct
